Applying corrections
