+++ inginious/frontend/plugins/contact_page/pages/contact_page.py
@@ -9,42 +9,23 @@ import web
 import requests
 import json
 from inginious.frontend.pages.utils import INGIniousPage
-from inginious.frontend.plugins.contact_page.pages.api.slack_url_error import SlackURLError
-
-subject_text = {
-    "subject-comment": "Comentario o reporte de un problema",
-    "subject-new-course": "Peticion para crear un nuevo curso",
-}
-URL_channel = {
-    "subject-comment": "",
-    "subject-new-course": "",
-}
-subject_new_course_id = "subject-new-course"
+from .constants import URL_channel, SUBJECT_NEW_COURSE_ID
 
 base_renderer_path = "frontend/plugins/contact_page/pages/templates"
 
 
-def set_url_channel(main_message_channel, new_course_channel):
-    """ Define the URL directions where do the request """
-    global URL_channel
-    if main_message_channel != "":
-        URL_channel["subject-comment"] = main_message_channel
-        if new_course_channel != "":
-            URL_channel["subject-new-course"] = new_course_channel
-        else:
-            URL_channel["subject-new-course"] = main_message_channel
-    else:
-        raise SlackURLError("The main slack's URL is empty")
-
-
-def create_payload(data):
-    """ Create the payload to be send """
-    blocks = [create_slack_text_block("Asunto", subject_text[data["subject_id"]]),
+def generate_slack_request_payload(data):
+    """ Create the payload to be sent """
+    subject_text = {
+        "subject-comment": "Report a problem or make a comment",
+        "subject-new-course": "Request to create a new course",
+    }
+    blocks = [create_slack_text_block("Subject", subject_text[data["subject_id"]]),
               create_slack_text_block("Email", data["email"]),
-              create_slack_text_block("Nombre", data["name"])]
+              create_slack_text_block("Name", data["name"])]
     if is_new_course_request(data["subject_id"]):
-        blocks.append(create_slack_text_block("Nombre del curso", data["courseName"]))
-    blocks.append(create_slack_text_block("Comentario", data["textarea"]))
+        blocks.append(create_slack_text_block("Course Name", data["courseName"]))
+    blocks.append(create_slack_text_block("Comments", data["textarea"]))
     return json.dumps({"blocks": blocks})
 
 
@@ -70,11 +51,12 @@ def send_request_to_slack(subject_id, payload):
 
 def is_new_course_request(subject_id):
     """  """
-    return subject_id == subject_new_course_id
+    return subject_id == SUBJECT_NEW_COURSE_ID
 
 
 class ContactPage(INGIniousPage):
     """ Contact page """
+
     def GET(self):
         """ Get request. Return the contact page """
         return self.template_helper.get_custom_renderer(base_renderer_path).contact_page()
@@ -83,6 +65,6 @@ class ContactPage(INGIniousPage):
         """ Post request. send a message to slack """
         data = web.input()
         subject_id = data["subject_id"]
-        payload = create_payload(data)
+        payload = generate_slack_request_payload(data)
         send_request_to_slack(subject_id, payload)
         return self.template_helper.get_custom_renderer(base_renderer_path).contact_page()
